requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
SQLAlchemy==2.0.34
psycopg[binary]==3.2.3
pydantic==2.9.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
starlette==0.38.5
alembic==1.13.2
requests==2.31.0
